043: solved w/ algorithms lib for permutations
This skips over the work done for 041 in implementing QuickPerm by
pulling in the swift algorithms library.

Also fixes a bug in the digitsToInt func that build it backwards, which
apparently did not matter for the last problem since it was just
checking the entire number for primality, but did in this one since we
are checking subsets of the digits where order matters.