050: add basic brute-force solution
It does not terminate under the "one minute" guideline, but does find a
solution within that time frame. As the liklihood of a solution falls
off quickly as we stop considering earlier numbers, so grabbing the last
appearing early solution works well enough!

It seems plausible that one could determine that a solution is
impossible after a certain point, based on solutions already seen, with
a proof... Though the nature of prime sequences may make that tricky?
Might have to be computational rather than emperical. Anyhow, what's
here is good enough.