044: iterative solution implemented
This was originally hobbled by a very slow implementation of perf_square
that used iteration, things fell into place quickly once that was
resolved. I also originally had the loop for "j" much higher than it
needed to be, slowing down the execution. This has been reduced now that
the solution is known for faster subsequent iteration.

Used a Makefile here instead of Cargo just to see what that would be
like.